+++ service/frontend/twittch_handler.go
@@ -13,12 +13,10 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-//BiliBiliMessage message handler
-func BiliBiliMessage(s *discordgo.Session, m *discordgo.MessageCreate) {
-	Prefix := configfile.BotPrefix.Bilibili
+func TwitchMessage(s *discordgo.Session, m *discordgo.MessageCreate) {
+	Prefix := configfile.BotPrefix.Twitch
 	m.Content = strings.ToLower(m.Content)
 	if strings.HasPrefix(m.Content, Prefix) {
-		loc, _ := time.LoadLocation("Asia/Shanghai") /*Use CST*/
 		CommandArray := strings.Split(m.Content, " ")
 		if len(CommandArray) > 1 {
 			Payload := strings.Split(strings.TrimSpace(CommandArray[1]), ",")
@@ -31,28 +29,36 @@ func BiliBiliMessage(s *discordgo.Session, m *discordgo.MessageCreate) {
 							s.ChannelMessageSend(m.ChannelID, "`"+FindGroupArry+"`,Name of Vtuber Group or Vtuber Name was not found")
 							return
 						} else {
-							LiveBili := database.BilGet(0, Member.ID, config.LiveStatus)
+							LiveTwitch := database.TwitchGet(0, Member.ID, config.LiveStatus)
 							FixName := engine.FixName(Member.EnName, Member.JpName)
-							if LiveBili != nil {
+							if LiveTwitch != nil {
 								Color, err := engine.GetColor(config.TmpDir, m.Author.AvatarURL("128"))
 								if err != nil {
 									log.Error(err)
 								}
-								for _, LiveData := range LiveBili {
+								for _, LiveData := range LiveTwitch {
+									FanBase := "simps"
+									loc := engine.Zawarudo(Member.Region)
 									diff := time.Now().In(loc).Sub(LiveData.Schedul.In(loc))
 									view, err := strconv.Atoi(LiveData.Viewers)
 									if err != nil {
 										log.Error(err)
 									}
+
+									if Member.Fanbase != "" {
+										FanBase = Member.Fanbase
+									}
+
 									_, err = s.ChannelMessageSendEmbed(m.ChannelID, engine.NewEmbed().
 										SetTitle(FixName).
 										SetAuthor(m.Author.Username, m.Author.AvatarURL("128")).
-										SetDescription(LiveData.Desc).
-										SetThumbnail(Member.BiliBiliAvatar).
+										SetThumbnail(Member.TwitchAvatar).
 										SetImage(LiveData.Thumb).
-										SetURL("https://live.bilibili.com/"+strconv.Itoa(Member.BiliRoomID)).
+										SetURL("https://twitch.tv/"+Member.TwitchName).
 										AddField("Start live", durafmt.Parse(diff).LimitFirstN(2).String()+" Ago").
-										AddField("Online", engine.NearestThousandFormat(float64(view))).
+										AddField("Viewers", engine.NearestThousandFormat(float64(view))+" "+FanBase).
+										InlineAllFields().
+										AddField("Game", LiveData.Game).
 										SetColor(Color).
 										SetFooter(LiveData.Schedul.In(loc).Format(time.RFC822)).MessageEmbed)
 									if err != nil {
@@ -70,30 +76,38 @@ func BiliBiliMessage(s *discordgo.Session, m *discordgo.MessageCreate) {
 							}
 						}
 					} else {
-						LiveBili := database.BilGet(VTuberGroup.ID, 0, config.LiveStatus)
-						if LiveBili != nil {
+						TwitchLive := database.TwitchGet(VTuberGroup.ID, 0, config.LiveStatus)
+						if TwitchLive != nil {
 							Color, err := engine.GetColor(config.TmpDir, m.Author.AvatarURL("128"))
 							if err != nil {
 								log.Error(err)
 							}
 
-							for _, LiveData := range LiveBili {
+							for _, LiveData := range TwitchLive {
+								FanBase := "simps"
 								LiveData.AddMember(FindVtuber(LiveData.Member.ID))
+								loc := engine.Zawarudo(LiveData.Member.Region)
 								FixName := engine.FixName(LiveData.Member.EnName, LiveData.Member.JpName)
 								diff := time.Now().In(loc).Sub(LiveData.Schedul.In(loc))
 								view, err := strconv.Atoi(LiveData.Viewers)
 								if err != nil {
 									log.Error(err)
 								}
+
+								if LiveData.Member.Fanbase != "" {
+									FanBase = LiveData.Member.Fanbase
+								}
+
 								_, err = s.ChannelMessageSendEmbed(m.ChannelID, engine.NewEmbed().
 									SetTitle(FixName).
 									SetAuthor(m.Author.Username, m.Author.AvatarURL("128")).
-									SetDescription(LiveData.Desc).
-									SetThumbnail(LiveData.Member.BiliBiliAvatar).
+									SetThumbnail(LiveData.Member.TwitchAvatar).
 									SetImage(LiveData.Thumb).
-									SetURL("https://live.bilibili.com/"+strconv.Itoa(LiveData.Member.BiliRoomID)).
+									SetURL("https://twitch.tv/"+LiveData.Member.TwitchName).
 									AddField("Start live", durafmt.Parse(diff).LimitFirstN(2).String()+" Ago").
-									AddField("Online", engine.NearestThousandFormat(float64(view))).
+									AddField("Viewers", engine.NearestThousandFormat(float64(view))+" "+FanBase).
+									InlineAllFields().
+									AddField("Game", LiveData.Game).
 									SetColor(Color).
 									SetFooter(LiveData.Schedul.In(loc).Format(time.RFC822)).MessageEmbed)
 								if err != nil {
@@ -121,29 +135,37 @@ func BiliBiliMessage(s *discordgo.Session, m *discordgo.MessageCreate) {
 							s.ChannelMessageSend(m.ChannelID, "`"+FindGroupArry+"`,Name of Vtuber Group or Vtuber Name was not found")
 							return
 						} else {
-							LiveBili := database.BilGet(0, Member.ID, config.PastStatus)
+							TwitchLive := database.TwitchGet(0, Member.ID, config.PastStatus)
 							FixName := engine.FixName(Member.JpName, Member.EnName)
-							if LiveBili != nil {
+							if TwitchLive != nil {
 								Color, err := engine.GetColor(config.TmpDir, m.Author.AvatarURL("128"))
 								if err != nil {
 									log.Error(err)
 								}
 
-								for _, LiveData := range LiveBili {
+								for _, LiveData := range TwitchLive {
+									FanBase := "simps"
+									loc := engine.Zawarudo(Member.Region)
 									diff := LiveData.Schedul.In(loc).Sub(time.Now().In(loc))
 									view, err := strconv.Atoi(LiveData.Viewers)
 									if err != nil {
 										log.Error(err)
 									}
+
+									if Member.Fanbase != "" {
+										FanBase = Member.Fanbase
+									}
+
 									_, err = s.ChannelMessageSendEmbed(m.ChannelID, engine.NewEmbed().
 										SetTitle(FixName).
 										SetAuthor(m.Author.Username, m.Author.AvatarURL("128")).
-										SetDescription(LiveData.Desc).
-										SetThumbnail(Member.BiliBiliAvatar).
+										SetThumbnail(Member.TwitchAvatar).
 										SetImage(LiveData.Thumb).
-										SetURL("https://live.bilibili.com/"+strconv.Itoa(Member.BiliRoomID)).
+										SetURL("https://twitch.tv/"+Member.TwitchName).
 										AddField("Start live", durafmt.Parse(diff).LimitFirstN(2).String()+" Ago").
-										AddField("Online", engine.NearestThousandFormat(float64(view))).
+										AddField("Viewers", engine.NearestThousandFormat(float64(view))+" "+FanBase).
+										InlineAllFields().
+										AddField("Game", LiveData.Game).
 										SetColor(Color).
 										SetFooter(LiveData.Schedul.In(loc).Format(time.RFC822)).MessageEmbed)
 									if err != nil {
@@ -161,30 +183,39 @@ func BiliBiliMessage(s *discordgo.Session, m *discordgo.MessageCreate) {
 							}
 						}
 					} else {
-						LiveBili := database.BilGet(VTuberGroup.ID, 0, config.PastStatus)
-						if LiveBili != nil {
+						TwitchLive := database.TwitchGet(VTuberGroup.ID, 0, config.PastStatus)
+						if TwitchLive != nil {
 							Color, err := engine.GetColor(config.TmpDir, m.Author.AvatarURL("128"))
 							if err != nil {
 								log.Error(err)
 							}
 
-							for _, LiveData := range LiveBili {
+							for _, LiveData := range TwitchLive {
+								FanBase := "simps"
 								LiveData.AddMember(FindVtuber(LiveData.Member.ID))
+								loc := engine.Zawarudo(LiveData.Member.Region)
+
 								FixName := engine.FixName(LiveData.Member.EnName, LiveData.Member.JpName)
 								view, err := strconv.Atoi(LiveData.Viewers)
 								if err != nil {
 									log.Error(err)
 								}
+
+								if LiveData.Member.Fanbase != "" {
+									FanBase = LiveData.Member.Fanbase
+								}
+
 								diff := time.Now().In(loc).Sub(LiveData.Schedul.In(loc))
 								_, err = s.ChannelMessageSendEmbed(m.ChannelID, engine.NewEmbed().
 									SetTitle(FixName).
 									SetAuthor(m.Author.Username, m.Author.AvatarURL("128")).
-									SetDescription(LiveData.Desc).
-									SetThumbnail(LiveData.Member.BiliBiliAvatar).
+									SetThumbnail(LiveData.Member.TwitchAvatar).
 									SetImage(LiveData.Thumb).
-									SetURL("https://live.bilibili.com/"+strconv.Itoa(LiveData.Member.BiliRoomID)).
+									SetURL("https://twitch.tv/"+LiveData.Member.TwitchName).
 									AddField("Start live", durafmt.Parse(diff).LimitFirstN(2).String()+" Ago").
-									AddField("Online", engine.NearestThousandFormat(float64(view))).
+									AddField("Viewers", engine.NearestThousandFormat(float64(view))+" "+FanBase).
+									InlineAllFields().
+									AddField("Game", LiveData.Game).
 									SetColor(Color).
 									SetFooter(LiveData.Schedul.In(loc).Format(time.RFC822)).MessageEmbed)
 								if err != nil {
@@ -200,123 +231,6 @@ func BiliBiliMessage(s *discordgo.Session, m *discordgo.MessageCreate) {
 								log.Error(err)
 							}
 						}
-					}
-				}
-			}
-		} else {
-			_, err := s.ChannelMessageSend(m.ChannelID, "Incomplete `"+Prefix+"` command")
-			if err != nil {
-				log.Error(err)
-			}
-		}
-	}
-}
-
-//BiliBiliSpace message hadler
-func BiliBiliSpace(s *discordgo.Session, m *discordgo.MessageCreate) {
-	m.Content = strings.ToLower(m.Content)
-	Prefix := "sp_" + configfile.BotPrefix.Bilibili
-	if strings.HasPrefix(m.Content, Prefix) {
-		loc, _ := time.LoadLocation("Asia/Shanghai") /*Use CST*/
-		Payload := m.Content[len(Prefix):]
-		if Payload != "" {
-			for _, FindGroupArry := range strings.Split(strings.TrimSpace(Payload), ",") {
-				VTuberGroup, err := FindGropName(FindGroupArry)
-				if err != nil {
-					Member := FindVtuber(FindGroupArry)
-					if Member.IsMemberNill() {
-						s.ChannelMessageSend(m.ChannelID, "`"+FindGroupArry+"`,Name of Vtuber Group or Vtuber Name was not found")
-						return
-					} else {
-						SpaceBili := database.SpaceGet(0, Member.ID)
-						FixName := engine.FixName(Member.EnName, Member.JpName)
-						if SpaceBili != nil {
-							Color, err := engine.GetColor(config.TmpDir, m.Author.AvatarURL("128"))
-							if err != nil {
-								log.Error(err)
-							}
-
-							for _, SpaceData := range SpaceBili {
-								diff := time.Now().In(loc).Sub(SpaceData.Schedul.In(loc))
-								view, err := strconv.Atoi(SpaceData.Viewers)
-								if err != nil {
-									log.Error(err)
-								}
-								_, err = s.ChannelMessageSendEmbed(m.ChannelID, engine.NewEmbed().
-									SetAuthor(m.Author.Username, m.Author.AvatarURL("128")).
-									SetTitle(FixName).
-									SetDescription(SpaceData.Title).
-									SetImage(SpaceData.Thumb).
-									SetThumbnail(Member.BiliBiliAvatar).
-									SetURL("https://www.bilibili.com/video/"+SpaceData.VideoID).
-									AddField("Type", SpaceData.Type).
-									AddField("Video uploaded", durafmt.Parse(diff).LimitFirstN(2).String()+" Ago").
-									AddField("Duration", SpaceData.Length).
-									AddField("Viewers now", engine.NearestThousandFormat(float64(view))).
-									SetFooter(SpaceData.Schedul.In(loc).Format(time.RFC822), config.BiliBiliIMG).
-									InlineAllFields().
-									SetColor(Color).MessageEmbed)
-								if err != nil {
-									log.Error(err)
-								}
-							}
-						} else {
-							_, err := s.ChannelMessageSendEmbed(m.ChannelID, engine.NewEmbed().
-								SetAuthor(m.Author.Username, m.Author.AvatarURL("128")).
-								SetDescription("It looks like `"+FixName+"` doesn't have a video in space.bilibili").
-								SetImage(config.WorryIMG).MessageEmbed)
-							if err != nil {
-								log.Error(err)
-							}
-							return
-						}
-					}
-					break
-				} else {
-					SpaceBili := database.SpaceGet(VTuberGroup.ID, 0)
-					if SpaceBili != nil {
-						Color, err := engine.GetColor(config.TmpDir, m.Author.AvatarURL("128"))
-						if err != nil {
-							log.Error(err)
-						}
-
-						for _, SpaceData := range SpaceBili {
-							SpaceData.AddMember(FindVtuber(SpaceData.Member.ID))
-
-							FixName := engine.FixName(SpaceData.Member.EnName, SpaceData.Member.JpName)
-							diff := time.Now().In(loc).Sub(SpaceData.Schedul.In(loc))
-							view, err := strconv.Atoi(SpaceData.Viewers)
-							if err != nil {
-								log.Error(err)
-							}
-							_, err = s.ChannelMessageSendEmbed(m.ChannelID, engine.NewEmbed().
-								SetAuthor(m.Author.Username, m.Author.AvatarURL("128")).
-								SetTitle(FixName).
-								SetDescription(SpaceData.Title).
-								SetImage(SpaceData.Thumb).
-								SetThumbnail(SpaceData.Member.BiliBiliAvatar).
-								SetURL("https://www.bilibili.com/video/"+SpaceData.VideoID).
-								AddField("Type", SpaceData.Type).
-								AddField("Video uploaded", durafmt.Parse(diff).LimitFirstN(2).String()+" Ago").
-								AddField("Duration", SpaceData.Length).
-								AddField("Viewers now", engine.NearestThousandFormat(float64(view))).
-								SetFooter(SpaceData.Schedul.In(loc).Format(time.RFC822), config.BiliBiliIMG).
-								InlineAllFields().
-								SetColor(Color).MessageEmbed)
-							if err != nil {
-								log.Error(err)
-							}
-
-						}
-					} else {
-						_, err := s.ChannelMessageSendEmbed(m.ChannelID, engine.NewEmbed().
-							SetAuthor(m.Author.Username, m.Author.AvatarURL("128")).
-							SetDescription("It looks like `"+VTuberGroup.GroupName+"` doesn't have a video in space.bilibili").
-							SetImage(config.WorryIMG).MessageEmbed)
-						if err != nil {
-							log.Error(err)
-						}
-						return
 					}
 				}
 			}